refactor: io 패키지 네이밍 view로 수정
