+++ src/main/java/RacingApplication.java
@@ -1,5 +1,5 @@
-import io.InputView;
-import io.OutputView;
+import view.InputView;
+import view.OutputView;
 
 public class RacingApplication {
