+++ src/main/java/view/OutputView.java
@@ -1,4 +1,4 @@
-package io;
+package view;
 
 import domain.Car;
