+++ src/test/java/CarRacingTest.java
@@ -1,19 +1,16 @@
 import common.exception.message.ExceptionMessage;
 import common.exception.model.ValidateException;
 import domain.*;
-import io.InputView;
-import io.OutputView;
+import view.InputView;
+import view.OutputView;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.List;
-import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
